requirements.txt
DrissionPage==4.1.0.9
colorama==0.4.6
python-dotenv==1.0.0
pyinstaller==6.3.0
requests==2.31.0
flet>=0.28.0
selenium==4.18.1
pyautogui==0.9.54
pyperclip==1.8.2
psutil==5.9.8
imaplib2==3.6
email-validator==2.1.0.post1
base64io==1.0.3
beautifulsoup4==4.12.2
pillow==10.2.0
pynput==1.7.6
uuid==1.30
datetime==5.4
traceback2==1.4.0
argparse==1.4.0
regex==2023.12.25